+++ src/geom/metric/mod.rs
@@ -15,6 +15,10 @@ crate::structural_mods! { // _reexports
         #[doc(inline)]
         pub use devela_base_core::{
             Distance, Extent, Position, Region, Stride,
+            // aliases
+            Distance1, Distance2, Distance3, Extent1, Extent2, Extent3,
+            Position1, Position2, Position3, Region1, Region2, Region3,
+            RegionS, RegionS1, RegionS2, RegionS3, Stride1, Stride2, Stride3,
         };
     }
 }